Moved unshared components out of shared
