+++ src/app/contributing/contributing.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-controbuting',
+  selector: 'app-contributing',
   templateUrl: './contributing.component.html',
   styleUrls: ['./contributing.component.scss']
 })
