+++ src/app/installation/installation.component.ts
