+++ src/app/references/reference/reference.component.html
