+++ src/app/references/reference/reference.component.scss
