+++ src/app/references/references.component.ts
